backend/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
py-postgresql==1.2.1
pyshp==2.1.2
Werkzeug==1.0.1
